add binary search tree

+++ Data_Structures_and_Algorithm_Analysis_in_C/2_algorithm_analysis/rapid_pow.py
@@ -5,7 +5,7 @@ def rapid_pow(x, n):
     """
     :type x: int.
 
-    :type n: int
+    :type n: float
     :rtype: int
     """
     if n == 0:
